Module Webdav_fs.Make

Parameters

module Pclock : Mirage_clock.PCLOCK

module Fs : Mirage_kv.RW

Signature

val (>>==) : ( 'a, 'b ) Stdlib.result Lwt.t -> ( 'a -> ( 'c, 'b ) Stdlib.result Lwt.t ) -> ( 'c, 'b ) Stdlib.result Lwt.t

type t = Fs.t

type error

type write_error

val basename : file_or_dir -> string

val create_file : dir -> string -> file

val dir_from_string : string -> dir

val file_from_string : string -> file

val from_string : t -> string -> ( file_or_dir, error ) Stdlib.result Lwt.t

val to_string : file_or_dir -> string

val parent : file_or_dir -> dir

val get_property_map : t -> file_or_dir -> Properties.t Lwt.t

val write_property_map : t -> file_or_dir -> Properties.t -> ( unit, write_error ) Stdlib.result Lwt.t

val size : t -> file -> ( int64, error ) Stdlib.result Lwt.t

val read : t -> file -> ( string * Properties.t, error ) Stdlib.result Lwt.t

val exists : t -> string -> bool Lwt.t

val dir_exists : t -> dir -> bool Lwt.t

val listdir : t -> dir -> ( file_or_dir list, error ) Stdlib.result Lwt.t

val mkdir : t -> dir -> Properties.t -> ( unit, write_error ) Stdlib.result Lwt.t

val write : t -> file -> string -> Properties.t -> ( unit, write_error ) Stdlib.result Lwt.t

be careful to call only in a batch, since it writes two files

val destroy : t -> file_or_dir -> ( unit, write_error ) Stdlib.result Lwt.t

be careful to call only in a batch, since it removes two files

val pp_error : error Fmt.t

val pp_write_error : write_error Fmt.t

val valid : t -> Webdav_config.config -> ( unit, [> `Msg of string ] ) Stdlib.result Lwt.t

val last_modified : t -> file_or_dir -> ( Ptime.t, error ) Stdlib.result Lwt.t

val etag : t -> file_or_dir -> ( string, error ) Stdlib.result Lwt.t

val batch : t -> ?retries:int -> ( t -> 'a Lwt.t ) -> 'a Lwt.t