Module type Webmachine.IO

The IO module signature abstracts over monadic futures library. It is a much reduced version of the module signature that appears in Cohttp, and as such is compatible with any module that conforms to Cohttp.S.IO.

type +'a t

The type of a blocking computation

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

The monadic bind operator for the type 'a t. m >>= f will pass the result of m to f, once the result is determined.

val return : 'a -> 'a t

return a creates a value of type 'a t that is already determined.